requirements.txt
cryptography>=42.0.2,<43
Pillow>=9.0.1,<10
numpy>=1.21.6,<2
xmltodict>=0.13.0,<1
requests>=2.32.3,<3
Flask>=3.0.3,<4
websockets>=12.0,<13
flask-sock>=0.7.0,<1
